+++ migrations/2025-12-17-033418_consumption-contest-table/up.sql
@@ -9,11 +9,8 @@ create table data.consumption_contests (
     food_emoji text not null,
     food text not null,
 
-    batting_team_mmolb_id text not null,
-    batting_team_emoji text not null,
-    batting_team_name text not null,
     batting_team_player_name text not null,
-    batting_team_score integer not null,
+    batting_team_total_consumed integer not null,
     batting_team_tokens integer not null,
     batting_team_prize_emoji text, -- null = batting team didn't win or tie
     batting_team_prize_name text, -- null = batting team didn't win or tie
@@ -21,11 +18,8 @@ create table data.consumption_contests (
     batting_team_prize_prefix text, -- null = batting team didn't win or tie, or won an item without a prefix
     batting_team_prize_suffix text, -- null = batting team didn't win or tie, or won an item without a suffix
 
-    defending_team_mmolb_id text not null,
-    defending_team_emoji text not null,
-    defending_team_name text not null,
     defending_team_player_name text not null,
-    defending_team_score integer not null,
+    defending_team_total_consumed integer not null,
     defending_team_tokens integer not null,
     defending_team_prize_emoji text, -- null = batting team didn't win or tie
     defending_team_prize_name text, -- null = batting team didn't win or tie